Minor refactor of internal components to reduce generated chunks
