+++ packages/hydrogen/src/foundation/index.tsx
@@ -1,3 +1,8 @@
-export * from './ServerPropsProvider';
+export {
+  ServerPropsProvider,
+  ServerPropsContext,
+  type ServerProps,
+  type ServerPropsContextValue,
+} from './ServerPropsProvider';
 export {useShop} from './useShop';
 export {useUrl} from './useUrl';
